File System Changes
While giving a workshop in Hong Kong, I discovered a key problem - all
that was making the app successful in quantifying data was passive
storage of environmental variables in the file. I stripped these,
switching from save() and load() to saveRDS() and load(RDS), with a
function designed to strip unnecessary environmental information from
the lm models in the cal. This radically reduced the size of the files,
from ~6.5 mb to ~0.5 mb. It also (seems) to have solved the problem of
growing cal files as the ‘create cal’ button was clicked twice - the
redundant environmental data was being stored essentially as introns in
DNA

I rewrote the global data prep functions as well, this was overlooked
earlier because they worked fine with the global variables on the same
data set that was used to build the cal, but failed on any other.