Garage Sale
Discovered that the validation results differed from those in the
summary table (which has now been reformatted to be clearer). Decided
to tie the preparation of the cal directly to the global functions used
to prep the validation data, so that there is no chance of error. This
has the bonus effect of eliminating lots of unnecessary code. There
seems to be something wrong with compton normalization, but everything
else is working well.