Rainforest
A new calibration type (Rainforest, 5) has been added. It is a forest regression, but in place of using other elemental lines as variables,  individual counts themselves are used.

Global functions have been added to adjust these variables. They compress the spectrra to 0.1 keV resolution, then run forerst on their intensiteis. Slope variable importance now shows a modified spectrum showing which parts of the spectrum inform the model.

This is a fundamentally different type of calibration, as it uses the whole spectrum. Early testing suggests it may be useful, but as far as I know this type of model is unknown in this kind of spectroscopy.