Model reorganize
I've edited the organization of models to run each variant, then toggle between them. This way the model summary better aligns with the runs the user sees. The rainforest and forest models have lower r2 values in the model summary than they do in the actual data - not sure why.