Model Memory
Added memory for machine learning models, as well as a silo button that prevents the model being run every time a parameter is changed. An addded bonus is that differend models will have different memory storage, allowing for some permanance. The negative is that the ui no longer reflects active settings if a new model is switched immediatly.