The Great Simplification
For the past week, I've largely re-written the data handling and memory use in CloudCal. This involved updating the way reactivity works in shiny under this app, resulting in a far simpler path from a to b (with a being the data and b being the model). This has a lot of corrolary benfits - models can only be run once (not accidently triggered two extra times by slow reactivity responses). shinyreactlog (which is now on by default) shows magnitudes-of-order less complexity for generating a single element's model.

In addition to this, spectra-only models for forest, neural networks, and XGBoost now have three new controls:
compression level
transformations
energy range

And all machine learning models, if using repeated cross-validation as a train control, now have user-selected repeats.

Because these changes required updating the CalTable headers, global functions have been deployed which can fill in old cals with these new headers if they are missing. This ultimatly solves the compatability problems that have arisen with the switch to a machine-learning framework.

Last, but not least, the cal prediction function has been moved to global.