Narrow and Wide
Added support for wide element lines at the element model level, with the ability to set a calibration-wide preference. Line preference is now in CalTable. To accomodate this change, line table generation has been made more efficient - so table generation is twice as fast but now there is also two times as much work. Well, it might be a little faster.

LineType should hopeuflly be updated throughout - it does work with a calRDS fix to handle NA values, but unclear if this could cause problems down the line.

Line definitions have been placed in a .rdata file, and do not need to be generated when loaded - this makes the overall app a little bit snappier.

Unecessary data is now thrown out of the app for faster downloads - old manuals, example .quant files, and example spectra.