Fano Factor
Added the index of dispersion (fano factor) to estimate error in measurement (em) and integrate it with the error of quantification(currently SE calculated from RMSE). The calculation is as follows:

eqm = (D x SE) + SE

where D is the index of dispersion and SE is the standard error.

The user can select the fano window to calculate the error of measurement, this defaults to 0.7 to 0.9 keV, which handles most spectra aquisition parameters.