Full Deconvolution
Added Least Squares Deconvolution as an option for calibration with prediction support. Also reconfigured prediction functions to require less processing if errors were selected.

"Plot Spectrum" button also has been removed, with a plots added by default.

Deconvolution occurs as an option for plotting, but will also happen anyway when lines are committed. Deconvoluted data is stored in "Deconvoluted" within a cal file. It will not be rerun if it is already present.